+++ CHANGES.md
@@ -82,6 +82,7 @@ Changes in CUPS v2.3.5
 - Removed support for the (long deprecated and unused) `KeepAliveTimeout`
   directive in `cupsd.conf` (Issue #5733)
 - Fixed `@IF(name)` handling in `cupsd.conf` (Issue #5918)
+- The scheduler now supports the "everywhere" model directly (Issue #5919)
 - Fixed `job-pages-per-set` value for duplex print jobs.